+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
+### Added
+
+- export `RouterInstance` class.
+- export `types`.
+- export `isUrlNavigatable` method.